Add files via upload
This small program doesn't require any external component. It's made to call it from your python program, but it works for testing standalone as well. It autodetects the reader disconnection, and also the reconnections. It's tested with more different 1D and 2D barcode scanners. You can modify the LUT for the different keyboard emulations if it needs.